Muteservice added

+++ Bot.Business/Models/MuteDto.cs
@@ -12,7 +12,7 @@ namespace Bot.Business.Models
         public int UserId { get; set; }
         public long ChatId { get; set; }
         public int MuteDuration { get; set; }
-        public int MuteTime { get; set; }
+        public long MuteTime { get; set; }
         public long MessageId { get; set; }
     }
 }
