+++ Bot.Data/Models/Mute.cs
@@ -11,7 +11,7 @@ namespace Bot.Data.Models
         public int UserId { get; set; }
         public long ChatId { get; set; }
         public int MuteDuration { get; set; }
-        public int MuteTime { get; set; }
+        public long MuteTime { get; set; }
         public long MessageId { get; set; }
     }
 }
